refactor: ThemeProvider 도입 및 spacing 정의
