+++ src/styles/theme.ts
@@ -1,0 +1,4 @@
+const theme = {
+  spacingUnit: 8,
+}
+export default theme;
